requirements.txt
absl-py==0.2.0
astor==0.6.2
atari-py==0.1.1
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
future==0.16.0
gast==0.2.0
grpcio==1.11.0
gym==0.10.5
h5py==2.7.1
html5lib==0.9999999
idna==2.6
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scipy==1.0.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
